+++ src/main/java/io/github/ricoapon/readableregex/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * All the public interfaces and classes.
+ */
+package io.github.ricoapon.readableregex;